Moved get_closest to the IntegrationDistance class
