+++ openquake/hazardlib/calc/filters.py
@@ -63,12 +63,12 @@ import logging
 import collections
 from contextlib import contextmanager
 import numpy
+from scipy.interpolate import interp1d
 try:
     import rtree
 except ImportError:
     rtree = None
 from openquake.baselib.python3compat import raise_
-from openquake.hazardlib import valid
 from openquake.hazardlib.site import FilteredSiteCollection
 from openquake.hazardlib.geo.utils import fix_lons_idl
 
@@ -122,6 +122,125 @@ def filter_sites_by_distance_to_rupture(rupture, integration_distance, sites):
     return sites.filter(jb_dist <= integration_distance)
 
 
+def getdefault(dic_with_default, key):
+    """
+    :param dic_with_default: a dictionary with a 'default' key
+    :param key: a key that may be present in the dictionary or not
+    :returns: the value associated to the key, or to 'default'
+    """
+    try:
+        return dic_with_default[key]
+    except KeyError:
+        return dic_with_default['default']
+
+
+def get_distances(rupture, mesh, param):
+    """
+    :param rupture: a rupture
+    :param mesh: a mesh of points
+    :param param: the kind of distance to compute (default rjb)
+    :returns: an array of distances from the given mesh
+    """
+    if param == 'rrup':
+        dist = rupture.surface.get_min_distance(mesh)
+    elif param == 'rx':
+        dist = rupture.surface.get_rx_distance(mesh)
+    elif param == 'ry0':
+        dist = rupture.surface.get_ry0_distance(mesh)
+    elif param == 'rjb':
+        dist = rupture.surface.get_joyner_boore_distance(mesh)
+    elif param == 'rhypo':
+        dist = rupture.hypocenter.distance_to_mesh(mesh)
+    elif param == 'repi':
+        dist = rupture.hypocenter.distance_to_mesh(mesh, with_depths=False)
+    elif param == 'rcdpp':
+        dist = rupture.get_cdppvalue(mesh)
+    elif param == 'azimuth':
+        dist = rupture.surface.get_azimuth(mesh)
+    else:
+        raise ValueError('Unknown distance measure %r' % param)
+    return dist
+
+
+class FarAwayRupture(Exception):
+    """Raised if the rupture is outside the maximum distance for all sites"""
+
+
+class IntegrationDistance(collections.Mapping):
+    """
+    Pickleable object wrapping a dictionary of integration distances per
+    tectonic region type. The integration distances can be scalars or
+    list of pairs (magnitude, distance). Here is an example using 'default'
+    as tectonic region type sot that the same values will be used for all
+    tectonic region types:
+
+    >>> maxdist = IntegrationDistance({'default': [
+    ...          (1, 10), (2, 20), (3, 30), (4, 40), (5, 100), (6, 200),
+    ...          (7, 400), (8, 800)]})
+    >>> maxdist('Some TRT', mag=5.5)
+    array(150.0)
+
+    It has also a method `.get_closest(sites, rupture)` returning the closest
+    sites to the rupture and their distances. The integration distance can be
+    missing if the sites have been already filtered (empty dictionary): in
+    that case the method returns all the sites and all the distances.
+    """
+    def __init__(self, dic):
+        self.dic = dic  # TRT -> float or list of pairs
+        self.magdist = {}  # TRT -> (magnitudes, distances)
+        for trt, value in dic.items():
+            if isinstance(value, list):  # assume a list of pairs (mag, dist)
+                value.sort()  # make sure the list is sorted by magnitude
+                self.magdist[trt] = zip(*value)
+            else:
+                self.dic[trt] = float(value)
+
+    def __call__(self, trt, mag=None):
+        value = getdefault(self.dic, trt)
+        if isinstance(value, float):  # scalar maximum distance
+            return value
+        elif mag is None:  # get the maximum magnitude distance
+            return value[-1][1]
+        elif not hasattr(self, 'interp'):
+            self.interp = {}  # function cache
+        try:
+            md = self.interp[trt]  # retrieve from the cache
+        except KeyError:  # fill the cache
+            magdist = getdefault(self.magdist, trt)
+            md = self.interp[trt] = interp1d(
+                *magdist, bounds_error=False, fill_value='extrapolate')
+        return md(mag)
+
+    def get_closest(self, sites, rupture, distance_type='rrup'):
+        """
+        :param sites: a (Filtered)SiteColletion
+        :param rupture: a rupture
+        :param distance_type: default 'rjb'
+        :returns: (close sites, close distances)
+        :raises: a FarAwayRupture exception if the rupture is far away
+        """
+        distances = get_distances(rupture, sites.mesh, distance_type)
+        if not self.dic:  # for sites already filtered
+            return sites, distances
+        mask = distances <= self(rupture.tectonic_region_type, rupture.mag)
+        if mask.any():
+            return sites.filter(mask), distances[mask]
+        else:
+            raise FarAwayRupture
+
+    def __getitem__(self, trt):
+        return self(trt)
+
+    def __iter__(self):
+        return iter(self.dic)
+
+    def __len__(self):
+        return len(self.dic)
+
+    def __repr__(self):
+        return repr(self.dic)
+
+
 class SourceFilter(object):
     """
     The SourceFilter uses the rtree library if available. The index is
@@ -149,12 +268,7 @@ class SourceFilter(object):
         by default True, i.e. try to use the rtree module if available
     """
     def __init__(self, sitecol, integration_distance, use_rtree=True):
-        if (isinstance(integration_distance, collections.Mapping) and not
-                isinstance(integration_distance, valid.IntegrationDistance)):
-            self.integration_distance = valid.IntegrationDistance(
-                integration_distance)
-        else:
-            self.integration_distance = integration_distance
+        self.integration_distance = IntegrationDistance(integration_distance)
         self.sitecol = sitecol
         self.use_rtree = use_rtree and rtree and (
             integration_distance and sitecol is not None)
@@ -234,4 +348,4 @@ class SourceFilter(object):
         return dict(integration_distance=self.integration_distance,
                     sitecol=self.sitecol, use_rtree=False)
 
-source_site_noop_filter = SourceFilter(None, None)
+source_site_noop_filter = SourceFilter(None, {})
