+++ openquake/hazardlib/gsim/base.py
@@ -35,6 +35,7 @@ import numpy
 
 from openquake.hazardlib import const
 from openquake.hazardlib import imt as imt_module
+from openquake.hazardlib.calc.filters import IntegrationDistance, get_distances
 from openquake.baselib.general import DeprecationWarning
 from openquake.baselib.python3compat import with_metaclass
 
@@ -109,49 +110,13 @@ class MetaGSIM(abc.ABCMeta):
             cls.instantiable = True
 
 
-def get_distances(rupture, mesh, param):
-    """
-    :param rupture: a rupture
-    :param mesh: a mesh of points
-    :param param: the kind of distance to compute (default rjb)
-    :returns: an array of distances from the given mesh
-    """
-    if param == 'rrup':
-        dist = rupture.surface.get_min_distance(mesh)
-    elif param == 'rx':
-        dist = rupture.surface.get_rx_distance(mesh)
-    elif param == 'ry0':
-        dist = rupture.surface.get_ry0_distance(mesh)
-    elif param == 'rjb':
-        dist = rupture.surface.get_joyner_boore_distance(mesh)
-    elif param == 'rhypo':
-        dist = rupture.hypocenter.distance_to_mesh(mesh)
-    elif param == 'repi':
-        dist = rupture.hypocenter.distance_to_mesh(mesh, with_depths=False)
-    elif param == 'rcdpp':
-        dist = rupture.get_cdppvalue(mesh)
-    elif param == 'azimuth':
-        dist = rupture.surface.get_azimuth(mesh)
-    else:
-        raise ValueError('Unknown distance measure %r' % param)
-    return dist
-
-
-class FarAwayRupture(Exception):
-    """Raised if the rupture is outside the maximum distance for all sites"""
-
-
 class ContextMaker(object):
     """
     A class to manage the creation of contexts for distances, sites, rupture.
-    It has also a method `.get_closest(sites, rupture)` returning the closest
-    sites to the rupture and their distances. The integration distance can be
-    None if the sites have been already filtered: in that case returns all the
-    sites and all the distances.
     """
     REQUIRES = ['DISTANCES', 'SITES_PARAMETERS', 'RUPTURE_PARAMETERS']
 
-    def __init__(self, gsims, maximum_distance=None):
+    def __init__(self, gsims, maximum_distance=IntegrationDistance({})):
         self.gsims = gsims
         self.maximum_distance = maximum_distance
         for req in self.REQUIRES:
@@ -290,28 +255,11 @@ class ContextMaker(object):
             and distance parameters) is unknown.
         """
         rctx = self.make_rupture_context(rupture)
-        sites, distances = self.get_closest(site_collection, rupture, 'rjb')
+        sites, distances = self.maximum_distance.get_closest(
+            site_collection, rupture, 'rjb')
         sctx = self.make_sites_context(sites)
         dctx = self.make_distances_context(sites, rupture, {'rjb': distances})
         return (sctx, rctx, dctx)
-
-    def get_closest(self, sites, rupture, distance_type='rrup'):
-        """
-        :param sites: a (Filtered)SiteColletion
-        :param rupture: a rupture
-        :param distance_type: default 'rjb'
-        :returns: (close sites, close distances)
-        :raises: a FarAwayRupture exception if the rupture is far away
-        """
-        distances = get_distances(rupture, sites.mesh, distance_type)
-        if self.maximum_distance is None:  # for sites already filtered
-            return sites, distances
-        mask = distances <= self.maximum_distance(
-            rupture.tectonic_region_type, rupture.mag)
-        if mask.any():
-            return sites.filter(mask), distances[mask]
-        else:
-            raise FarAwayRupture
 
 
 @functools.total_ordering
