+++ openquake/hazardlib/tests/calc/hazard_curve_new_test.py
@@ -92,7 +92,7 @@ class HazardCurvesTestCase01(unittest.TestCase):
                                               PMF([(0.7, 0), (0.3, 1)]),
                                               TRT.GEOTHERMAL)
         site = Site(Point(0.0, 0.0), 800, True, z1pt0=100., z2pt5=1.)
-        s_filter = SourceFilter(SiteCollection([site]), None)
+        s_filter = SourceFilter(SiteCollection([site]), {})
         self.sites = s_filter
         self.imtls = DictArray({'PGA': [0.01, 0.1, 0.3]})
         self.gsim_by_trt = {TRT.ACTIVE_SHALLOW_CRUST: SadighEtAl1997()}
@@ -191,7 +191,7 @@ class NankaiTestCase(unittest.TestCase):
         groups = nrml.parse(source_model, SourceConverter(
             investigation_time=50., rupture_mesh_spacing=2.))
         site = Site(Point(135.68, 35.68), 800, True, z1pt0=100., z2pt5=1.)
-        s_filter = SourceFilter(SiteCollection([site]), None)
+        s_filter = SourceFilter(SiteCollection([site]), {})
         imtls = DictArray({'PGV': [20, 40, 80]})
         gsim_by_trt = {'Subduction Interface': SiMidorikawa1999SInter()}
         hcurves = calc_hazard_curves(groups, s_filter, imtls, gsim_by_trt)
